Simplify profiles.
