--- src/xyz/_121407/schoolmanagement/entities/ProfileType.java
@@ -1,9 +1,0 @@
-package xyz._121407.schoolmanagement.entities;
-
-public enum ProfileType {
-    MATE_INFO,
-    MATE_INFO_INTENSIV_INFO,
-    BIO_CHIMIE,
-    STIINTE_SOCIALE,
-    FILOLOGIE
-}
