+++ src/xyz/_121407/schoolmanagement/gui/profiles/ProfileForm.java
@@ -1,0 +1,2 @@
+package xyz._121407.schoolmanagement.gui.profiles;public class ProfileForm {
+}
