+++ src/xyz/_121407/schoolmanagement/gui/profiles/ProfilesPanel.java
@@ -1,0 +1,2 @@
+package xyz._121407.schoolmanagement.gui.profiles;public class ProfilesPanel {
+}
