+++ src/xyz/_121407/schoolmanagement/services/EntityLoader.java
@@ -1,0 +1,2 @@
+package xyz._121407.schoolmanagement.services;public class EntityLoader {
+}
